+++ CHANGES.rst
@@ -24,6 +24,9 @@ https://github.com/zopefoundation/Zope/blob/4.0a6/CHANGES.rst
   (`tempstorage#8 <https://github.com/zopefoundation/tempstorage/issues/8>`_)
   (`tempstorage#12 <https://github.com/zopefoundation/tempstorage/issues/12>`_)
 
+- Reuse ``zope.publisher.http.splitport`` instead of defining our own
+  (`#683 <https://github.com/zopefoundation/Zope/issues/683>`_)
+
 
 4.1.1 (2019-07-02)
 ------------------